[azdo] fix signing tasks (#16)
It appears that the .yaml template I use for signing expects at least
a `SignList.xml` now:

    Get-Content : Cannot find path 'C:\A\1\_temp\artifact-signing\unsigned\nuget-unsigned\SignList.xml' because it does not exist.

This repo is unique because there are not actually any `.dll` files to
sign. The NuGet package is only `.targets` files.

For now, I've added a `SignList.xml` with no items. If I ever add a
`.dll` file one day, I'll have an example to follow.